requirements.txt
# Core Framework Dependencies
langchain-core>=0.1.0,<1.0.0
langchain-openai>=0.1.0,<1.0.0
pydantic>=2.0.0,<3.0.0
flask>=2.0.0,<3.0.0

# Database and Data Processing
pymysql>=1.0.0,<2.0.0
pandas>=1.5.0,<3.0.0

# Machine Learning and Embeddings
scikit-learn>=1.3.0,<2.0.0
sentence-transformers>=2.2.0,<3.0.0
FlagEmbedding>=1.2.0,<2.0.0
transformers>=4.30.0,<5.0.0
torch>=2.0.0,<3.0.0
faiss-cpu>=1.7.0,<2.0.0
# Note: Use faiss-gpu if GPU support is needed

# DataHub Integration
acryl-datahub>=0.10.0,<1.0.0

# Utility Libraries
cachetools>=5.0.0
python-dotenv>=1.0.0

# Note: The following are built-in Python modules (no installation needed):
# - asyncio (built-in since Python 3.4)
# - concurrent.futures (built-in since Python 3.2)
# - argparse (built-in since Python 2.7)
# - logging (built-in)
# - json, re, os, time, threading, etc. (all built-in)

# Optional Dependencies (uncomment if needed)
# numpy>=1.24.0          # For advanced numerical operations
# matplotlib>=3.7.0      # For data visualization
# jupyter>=1.0.0         # For notebook development
# pytest>=7.0.0          # For unit testing
# black>=23.0.0          # For code formatting
# flake8>=6.0.0          # For code linting 